fix: workflow not execute
