+++ src/popup/pages/Home.vue
@@ -216,6 +216,8 @@ async function initElementSelector() {
     await browser.tabs.sendMessage(tab.id, {
       type: 'automa-element-selector',
     });
+
+    window.close();
   } catch (error) {
     if (error.message.includes('Could not establish connection.')) {
       await browser.tabs.executeScript(tab.id, {
